+++ docs/implementors/core/convert/trait.AsRef.js
@@ -1,3 +1,0 @@
-(function() {var implementors = {};
-implementors["my_iot"] = [{"text":"impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.AsRef.html\" title=\"trait core::convert::AsRef\">AsRef</a>&lt;<a class=\"enum\" href=\"my_iot/core/value/enum.Value.html\" title=\"enum my_iot::core::value::Value\">Value</a>&gt; for <a class=\"enum\" href=\"my_iot/core/value/enum.Value.html\" title=\"enum my_iot::core::value::Value\">Value</a>","synthetic":false,"types":["my_iot::core::value::Value"]}];
-if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()