+++ docs/implementors/core/default/trait.Default.js
@@ -1,3 +1,0 @@
-(function() {var implementors = {};
-implementors["my_iot"] = [{"text":"impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/default/trait.Default.html\" title=\"trait core::default::Default\">Default</a> for <a class=\"struct\" href=\"my_iot/settings/struct.DashboardSettings.html\" title=\"struct my_iot::settings::DashboardSettings\">DashboardSettings</a>","synthetic":false,"types":["my_iot::settings::DashboardSettings"]}];
-if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()