+++ docs/my_iot/core/sidebar-items.js
@@ -1,1 +1,0 @@
-initSidebarItems({"mod":[["bus","Message bus"],["message","Describes a sensor reading and related structures."],["persistence","Database interface."],["services","Implements generic `Service` trait."],["value","Implements sensor reading value."]]});