+++ docs/my_iot/core/value/sidebar-items.js
@@ -1,1 +1,0 @@
-initSidebarItems({"enum":[["PointOfTheCompass","Points of the compass. Provides the common aliases in English and Dutch."],["Value","Sensor reading value."]]});