+++ docs/my_iot/services/clock/sidebar-items.js
@@ -1,1 +1,0 @@
-initSidebarItems({"fn":[["default_interval_millis",""]],"struct":[["Clock",""]]});