+++ docs/my_iot/services/prelude/sidebar-items.js
@@ -1,1 +1,0 @@
-initSidebarItems({"constant":[["USER_AGENT","`User-Agent` header used for all outcoming HTTP requests."]],"fn":[["client_builder",""],["default_client",""]],"struct":[["Client","A `Client` to make Requests with."]]});